requirements.txt
langchain
wikipedia
langchain-community
python-dotenv
pydantic
duckduckgo-search
langchain_huggingface
streamlit
langchain-nvidia-ai-endpoints
langchain-ollama
langchain-chroma